change readOffloadBytes/readOffloadError from BlobStoreBackedReadHandleImpl to BlobStoreBackedInputStreamImpl.

+++ managed-ledger/src/main/java/org/apache/bookkeeper/mledger/impl/LedgerOffloaderMXBeanImpl.java
@@ -236,7 +236,7 @@ public class LedgerOffloaderMXBeanImpl implements LedgerOffloaderMXBean {
         adder.add(1L);
     }
 
-    public void recordReadOffloadBytes(String topicName, int size) {
+    public void recordReadOffloadBytes(String topicName, long size) {
         if (topicName == null) {
             return;
         }
